+++ broadcast-transcoder/src/main/java/dk/statsbiblioteket/broadcasttranscoder/processors/MultistreamVideoTranscoderProcessor.java
@@ -68,7 +68,7 @@ public class MultistreamVideoTranscoderProcessor extends ProcessorChainElement {
         }
         try {
             long programLength = MetadataUtils.findProgramLengthMillis(request);
-            long timeout = programLength/context.getTranscodingTimeoutDivisor();
+            long timeout = (long) (programLength/context.getTranscodingTimeoutDivisor());
             logger.debug("Setting transcoding timeout for '" + context.getProgrampid() + "' to " + timeout + "ms" );
             request.setTranscoderCommand(clipperCommand);
             ExternalJobRunner.runClipperCommand(timeout, clipperCommand);
