+++ broadcast-transcoder/src/main/java/dk/statsbiblioteket/broadcasttranscoder/processors/UnistreamVideoTranscoderProcessor.java
@@ -42,9 +42,9 @@ public class UnistreamVideoTranscoderProcessor extends ProcessorChainElement {
             long timeout;
             if (request.getTimeoutMilliseconds() == 0l) {
                 long programLength = MetadataUtils.findProgramLengthMillis(request);
-                timeout = programLength/context.getTranscodingTimeoutDivisor();
+                timeout = (long) (programLength/context.getTranscodingTimeoutDivisor());
             } else if (request.getFfprobeDurationSeconds() != null) {
-                timeout = Math.round(request.getFfprobeDurationSeconds()*1000L)/context.getTranscodingTimeoutDivisor();
+                timeout = (long) (Math.round(request.getFfprobeDurationSeconds()*1000L)/context.getTranscodingTimeoutDivisor());
             } else {
                 timeout = request.getTimeoutMilliseconds();
             }
