+++ broadcast-transcoder/src/main/java/dk/statsbiblioteket/broadcasttranscoder/reklamefilm/ReklamefilmFileResolverProcessor.java
@@ -31,7 +31,7 @@ public class ReklamefilmFileResolverProcessor extends ProcessorChainElement {
         logger.info("Resolved " + pid + " to " + request.getClipperCommand());
         final long nominalDurationSeconds = 300L;
         request.setBitrate(mediafile.length()/nominalDurationSeconds);
-        request.setTimeoutMilliseconds(nominalDurationSeconds*1000L/context.getTranscodingTimeoutDivisor());
+        request.setTimeoutMilliseconds((long) (nominalDurationSeconds*1000L/context.getTranscodingTimeoutDivisor()));
         request.setFileFormat(FileFormatEnum.MPEG_PS);
         TranscodeRequest.FileClip clip = new TranscodeRequest.FileClip("'" + mediafile.getAbsolutePath() + "'");
         List<TranscodeRequest.FileClip> clips = new ArrayList<TranscodeRequest.FileClip>();
